web_server/requirements/common.txt
django~=4.0.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.1
django-redis~=5.4.0
channels ~=4.1.0
channels-redis ~=4.2.0